requirements.txt
# --- Data Extraction & APIs ---
requests>=2.31.0          # For scraping Competitor Pricing & Weather APIs
pandas>=2.0.0             # For small-data manipulation before Spark
boto3>=1.28.0             # For talking to AWS S3

# --- Snowflake Connectivity ---
snowflake-connector-python>=3.6.0
snowflake-sqlalchemy>=1.5.0
apache-airflow-providers-snowflake>=5.0.0  # Allows Airflow to trigger Snowflake

# --- Utilities ---
python-dotenv>=1.0.0      # To read your .env file
pyspark==3.5.1            # Matches your Docker image version

apache-airflow-providers-snowflake